+++ docs/api/axios.api.md
@@ -6,8 +6,37 @@
 import { AxiosInstance } from "axios";
 import { AxiosRequestConfig } from "axios";
 import { AxiosResponse } from "axios";
+import { ComputedRef } from "@vue/runtime-core";
 import { PromiseResultFactory } from "vue-composable";
-import { Ref } from "@vue/composition-api";
+import { Ref } from "@vue/runtime-core";
+
+// @public (undocumented)
+export function makeAxios(
+  client: AxiosInstance,
+  throwException?: boolean
+): {
+  client: AxiosInstance;
+  data: import("@vue/reactivity").ComputedRef<any>;
+  status: import("@vue/reactivity").ComputedRef<number>;
+  statusText: import("@vue/reactivity").ComputedRef<string>;
+  cancel: (message?: string | undefined) => void;
+  isCancelled:
+    | import("@vue/reactivity").Ref<false>
+    | import("@vue/reactivity").Ref<true>;
+  cancelledMessage:
+    | import("@vue/reactivity").Ref<undefined>
+    | import("@vue/reactivity").Ref<null>
+    | import("@vue/reactivity").Ref<string>;
+  exec: (
+    request: string | AxiosRequestConfig
+  ) => Promise<AxiosResponse<any> | undefined>;
+  promise: import("@vue/reactivity").Ref<
+    Promise<AxiosResponse<any>> | undefined
+  >;
+  result: import("@vue/reactivity").Ref<AxiosResponse<any> | null>;
+  loading: import("@vue/reactivity").Ref<boolean>;
+  error: import("@vue/reactivity").Ref<any>;
+};
 
 // Warning: (ae-forgotten-export) The symbol "AxiosReturn" needs to be exported by the entry point index.d.ts
 //
